requirements.txt
asgiref==3.6.0
dj-database-url==2.2.0
Django==4.2
django-cors-headers==3.14.0
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==23.0.0
packaging==24.1
psycopg2==2.9.9
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.12.2
whitenoise==6.7.0
